+++ include/backend/cpu/ops/initializer.h
@@ -21,22 +21,22 @@ struct RandomImpl;
 
 template <>
 struct ZerosImpl<Device::CPU> {
-    static Tensor execute(const std::vector<int>& shape, DataType dtype);
+    static Tensor execute(const std::vector<int64_t>& shape, DataType dtype);
 };
 
 template <>
 struct OnesImpl<Device::CPU> {
-    static Tensor execute(const std::vector<int>& shape, DataType dtype);
+    static Tensor execute(const std::vector<int64_t>& shape, DataType dtype);
 };
 
 template <>
 struct FillImpl<Device::CPU> {
-    static Tensor execute(const std::vector<int>& shape, DataType dtype, float value);
+    static Tensor execute(const std::vector<int64_t>& shape, DataType dtype, float value);
 };
 
 template <>
 struct RandomImpl<Device::CPU> {
-    static Tensor execute(const std::vector<int>& shape, DataType dtype, float min, float max);
+    static Tensor execute(const std::vector<int64_t>& shape, DataType dtype, float min, float max);
 };
 
 extern template struct ZerosImpl<Device::CPU>;
